[Feat] 평가 생성 controller 메소드 구현

+++ src/main/java/com/allclear/tastytrack/domain/user/entity/User.java
@@ -1,6 +1,7 @@
 package com.allclear.tastytrack.domain.user.entity;
 
 import com.allclear.tastytrack.domain.user.dto.UserInfo;
+
 import jakarta.persistence.*;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
@@ -16,7 +17,7 @@ public class User {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
+    private int id;
 
     @Column(nullable = false)
     private String username;
